more generics...
git-svn-id: https://ho1.svn.sourceforge.net/svnroot/ho1/trunk@524 b8de521a-fa13-0410-8113-fc7ffabbd59a

+++ Core/src/java/de/hattrickorganizer/credits/VAPSchild.java
@@ -73,7 +73,7 @@ public class VAPSchild implements DynamischesObjekt {
 
         //Transparenz fadeout
         if (gesamtZeit > 4500) {
-            float alphawert = Math.abs(((float) (gesamtZeit - 4500.0f) / 500.0f) - 1.0f);
+            float alphawert = Math.abs(((gesamtZeit - 4500.0f) / 500.0f) - 1.0f);
 
             if (alphawert > 1.0f) {
                 alphawert = 1.0f;
